// src/types/AnimeDetails.ts
export interface AnimeDetails {
  malId: number;
  koreanName: string;
  japanesesName: string;
  productionCompany: string;
  genre: string;
  animelistUrl: string;
  imageUrl: string;
  largeImageUrl: string;
  trailerUrl: string;
  score: number;
  rating: string;
  source: string;
  themes?: string;
  special?: string;
}